pyproject.toml
[project]
name = "quantization-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.18.55",
    "baostock>=0.9.1",
    "celery>=5.6.3",
    "dependency-injector>=4.49.0",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langgraph>=1.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
